Fix rfft dtype error on CPU with fp16 model
torch.fft.rfft does not support float16 on CPU. Cast to float32
before rfft and cast the result back to the original dtype so the
rest of the forward pass stays consistent.

+++ src/dl_models/severity/cbramod_backbone.py
@@ -71,8 +71,8 @@ class PatchEmbedding(nn.Module):
         patch_emb = patch_emb.permute(0, 2, 1, 3).contiguous().view(bz, ch_num, patch_num, self.d_model)
 
         mask_x = mask_x.contiguous().view(bz * ch_num * patch_num, patch_size)
-        spectral = torch.fft.rfft(mask_x, dim=-1, norm='forward')
-        spectral = torch.abs(spectral).contiguous().view(bz, ch_num, patch_num, 101)
+        spectral = torch.fft.rfft(mask_x.float(), dim=-1, norm='forward')
+        spectral = torch.abs(spectral).to(mask_x.dtype).contiguous().view(bz, ch_num, patch_num, 101)
         spectral_emb = self.spectral_proj(spectral)
 
         patch_emb = patch_emb + spectral_emb
